+++ test/unit/image.js
@@ -73,16 +73,12 @@
   }
 
   function _createImageObject(width, height, callback, options) {
+    options = options || {};
     var elImage = _createImageElement();
     setSrc(elImage, IMG_SRC, function() {
-      if (width !== elImage.width || height !== elImage.height) {
-        elImage.width = width;
-        elImage.height = height;
-        callback(new fabric.Image(elImage, options));
-      }
-      else {
-        callback(new fabric.Image(elImage, options));
-      }
+      options.width = width;
+      options.height = height;
+      callback(new fabric.Image(elImage, options));
     });
   }
 
@@ -656,8 +652,25 @@
       var data1 = image.toDataURL();
       var data2 = image.toDataURL();
       var data3 = image.toDataURL();
-      assert.equal(data1, data2, 'dataurl does not change 1');
-      assert.equal(data1, data3, 'dataurl does not change 2');
+      assert.ok(data1 === data2, 'dataurl does not change 1');
+      assert.ok(data1 === data3, 'dataurl does not change 2');
+      done();
+    });
+  });
+
+  QUnit.test('apply filters run isNeutralState implementation of filters', function(assert) {
+    var done = assert.async();
+    createImageObject(function(image) {
+      var run = false;
+      image.dirty = false;
+      var filter = new fabric.Image.filters.Brightness();
+      image.filters = [filter];
+      filter.isNeutralState = function() {
+        run = true;
+      };
+      assert.equal(run, false, 'isNeutralState did not run yet');
+      image.applyFilters();
+      assert.equal(run, true, 'isNeutralState did run');
       done();
     });
   });
@@ -669,6 +682,50 @@
       assert.equal(image.dirty, false, 'false apply filter dirty is false');
       image.applyFilters();
       assert.equal(image.dirty, false, 'After apply filter dirty is true');
+      done();
+    });
+  });
+
+  QUnit.test('apply filters reset _element and _filteredEl', function(assert) {
+    var done = assert.async();
+    createImageObject(function(image) {
+      var contrast = new fabric.Image.filters.Contrast({ contrast: 0.5 });
+      image.applyFilters();
+      var element = image._element;
+      var filtered = image._filteredEl;
+      image.filters = [contrast];
+      image.applyFilters();
+      assert.notEqual(image._element, element, 'image element has changed');
+      assert.notEqual(image._filteredEl, filtered, 'image _filteredEl element has changed');
+      assert.equal(image._element, image._filteredEl, 'after filtering elements are the same');
+      done();
+    });
+  });
+
+  QUnit.test('apply filters and resize filter', function(assert) {
+    var done = assert.async();
+    createImageObject(function(image) {
+      var contrast = new fabric.Image.filters.Contrast({ contrast: 0.5 });
+      var resizeFilter = new fabric.Image.filters.Resize();
+      image.filters = [contrast];
+      image.resizeFilter = resizeFilter;
+      var element = image._element;
+      var filtered = image._filteredEl;
+      image.scaleX = 0.4;
+      image.scaleY = 0.4;
+      image.applyFilters();
+      assert.notEqual(image._element, element, 'image element has changed');
+      assert.notEqual(image._filteredEl, filtered, 'image _filteredEl element has changed');
+      assert.equal(image._element, image._filteredEl, 'after filtering elements are the same');
+      image.applyResizeFilters();
+      assert.notEqual(image._element, image._filteredEl, 'after resizing the 2 elements differ');
+      assert.equal(image._lastScaleX, image.scaleX, 'after resizing we know how much we scaled');
+      assert.equal(image._lastScaleY, image.scaleY, 'after resizing we know how much we scaled');
+      image.applyFilters();
+      assert.equal(image._element, image._filteredEl, 'after filters again the elements changed');
+      assert.equal(image._lastScaleX, 1, 'lastScale X is reset');
+      assert.equal(image._lastScaleY, 1, 'lastScale Y is reset');
+      assert.equal(image._needsResize(), true, 'resizing is needed again');
       done();
     });
   });